requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
Django==3.2
django-elasticsearch-dsl==7.2.1
django-elasticsearch-dsl-drf==0.22.2
django-nine==0.2.5
djangorestframework==3.12.4
elasticsearch==7.16.1
elasticsearch-dsl==7.4.0
idna==3.3
psycopg2-binary==2.9.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
